changes in libgcrypt for lattice-based signatures

+++ libgcrypt/src/sexp.c
@@ -81,7 +81,7 @@
 
  */
 
-typedef unsigned short DATALEN;
+typedef unsigned long DATALEN;
 
 struct gcry_sexp
 {
